+++ lib/castle/configuration.rb
@@ -3,16 +3,33 @@
 module Castle
   # manages configuration variables
   class Configuration
-    SUPPORTED = %i[source_header request_timeout api_secret api_endpoint].freeze
     REQUEST_TIMEOUT = 30.0
     API_ENDPOINT = 'https://api.castle.io/v1'
+    WHITELISTED = [
+      'User-Agent',
+      'Accept-Language',
+      'Accept-Encoding',
+      'Accept-Charset',
+      'Accept',
+      'Accept-Datetime',
+      'X-Forwarded-For',
+      'Forwarded',
+      'X-Forwarded',
+      'X-Real-IP',
+      'REMOTE_ADDR'
+    ].freeze
+
+    BLACKLISTED = ['HTTP_COOKIE'].freeze
 
     attr_accessor :request_timeout, :source_header
-    attr_reader :api_secret, :api_endpoint
+    attr_reader :api_secret, :api_endpoint, :whitelisted, :blacklisted
 
     def initialize
+      @formatter = Castle::HeaderFormatter.new
       @request_timeout = REQUEST_TIMEOUT
       self.api_endpoint = API_ENDPOINT
+      self.whitelisted = WHITELISTED
+      self.blacklisted = BLACKLISTED
       self.api_secret = ''
     end
 
@@ -24,6 +41,14 @@ module Castle
 
     def api_secret=(value)
       @api_secret = ENV.fetch('CASTLE_API_SECRET', value).to_s
+    end
+
+    def whitelisted=(value)
+      @whitelisted = value ? value.map { |header| @formatter.call(header) } : []
+    end
+
+    def blacklisted=(value)
+      @blacklisted = value ? value.map { |header| @formatter.call(header) } : []
     end
 
     private